+++ ios/Runner/GeneratedPluginRegistrant.m
@@ -3,12 +3,10 @@
 //
 
 #import "GeneratedPluginRegistrant.h"
-#import <sqflite/SqflitePlugin.h>
 
 @implementation GeneratedPluginRegistrant
 
 + (void)registerWithRegistry:(NSObject<FlutterPluginRegistry>*)registry {
-  [SqflitePlugin registerWithRegistrar:[registry registrarForPlugin:@"SqflitePlugin"]];
 }
 
 @end